gdt: 16-bit trap/iomap_base in the TSS; fix far jump label and stack_top export
trap and iomap_base were uint32_t, making struct tss_entry 108 bytes instead
of 104. The CPU reads iomap_base at the fixed offset 102, which fell inside
the widened trap field: it read 0, meaning "I/O bitmap at offset 0" instead
of "no bitmap", granting ring 3 the ports rather than denying them. Caught
from the TSS limit in the QEMU monitor (0x6b instead of 0x67).

Also fixes three typos: the far jump target was .reload:cs, the ltr operand
was missing a comma, and boot.s exported stacl_top.

+++ arch/x86/gdt.h
@@ -58,8 +58,10 @@ struct tss_entry {
     uint32_t esp, ebp, esi, edi;
     uint32_t es, cs, ss, ds, fs, gs;
     uint32_t ldt;
-    uint32_t trap;
-    uint32_t iomap_base;
+    //these two are 16 bit, not 32: the CPU reads iomap_base at the fixed
+    //offset 102, so a wrong width here silently moves it out of place.
+    uint16_t trap;
+    uint16_t iomap_base;
 } __attribute__((packed));
 
 void gdt_init(void);
